Split the code into a library and runner
