+++ bin/hub/gameroomMatchmaking.ts
@@ -2,10 +2,10 @@
  * Copied from the original bot hub, with permission.
  */
 
-import { logger } from "../api";
+import { logger } from "../../src/api";
 import promClient from "prom-client";
-import { API_Character } from "../apiCharacter";
-import { API_Connector } from "../apiConnector";
+import { API_Character } from "../../src/apiCharacter";
+import { API_Connector } from "../../src/apiConnector";
 
 // Metrics
 const metric_registered = new promClient.Counter({
