+++ bin/hub/logic/kidnappersGameRoom.ts
@@ -2,9 +2,9 @@
  * Copied from the original bot hub, with permission.
  */
 
-import { logger, BC_PermissionLevel } from "../../api";
-import type { API_Connector } from "../../apiConnector";
-import type { API_Character } from "../../apiCharacter";
+import { logger, BC_PermissionLevel } from "../../../src/api";
+import type { API_Connector } from "../../../src/apiConnector";
+import type { API_Character } from "../../../src/apiCharacter";
 
 import promClient from "prom-client";
 
@@ -13,8 +13,8 @@ import { AdministrationLogic } from "./administrationLogic";
 import { MatchmakingNotifier } from "../gameroomMatchmaking";
 
 import _ from "lodash";
-import { AssetGet } from "../../item";
-import { BC_Server_ChatRoomMessage } from "../../logicEvent";
+import { AssetGet } from "../../../src/item";
+import { BC_Server_ChatRoomMessage } from "../../../src/logicEvent";
 import { ConfigFile } from "../../config";
 
 /** INSTRUCTIONS TO ADD A NEW ROLE
