+++ bin/main.ts
@@ -13,7 +13,7 @@
  */
 
 import { KidnappersGameRoom } from "./hub/logic/kidnappersGameRoom";
-import { API_Connector } from "./apiConnector";
+import { API_Connector } from "../src/apiConnector";
 import { RoleplaychallengeGameRoom } from "./hub/logic/roleplaychallengeGameRoom";
 import { Dare } from "./games/dare";
 import { readFile } from "fs/promises";
@@ -144,3 +144,17 @@ export async function startBot(): Promise<RopeyBot> {
         game: config.game,
     };
 }
+
+async function main() {
+    const { game } = await startBot();
+
+    if (!game) {
+        console.error("No game specified!");
+        process.exit(1);
+    }
+}
+
+main().catch((e) => {
+    console.error(e);
+    process.exit(1);
+});
