+++ tsconfig.json
@@ -12,7 +12,8 @@
       "node_modules/bc-stubs/bc/NativeDeclarations/Typedef.d.ts",
       "node_modules/bc-stubs/bc/NativeDeclarations/Messages.d.ts",
       "node_modules/bc-stubs/bc/NativeDeclarations/Female3DCG_Types.d.ts",
-      "src/**/*"
+      "src/**/*",
+      "bin/**/*",
   ],
   "ts-node": {
     "files": true
